Revert "feat(CL): core swaprouter logic and tests (merge to main 3) (#3496)"
This reverts commit f9058fb6f6dae70ab421f70938a490dffc08a444.
