+++ x/gamm/pool-models/stableswap/msgs.go
@@ -5,7 +5,6 @@ import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
 	"github.com/osmosis-labs/osmosis/v13/x/gamm/types"
-	swaproutertypes "github.com/osmosis-labs/osmosis/v13/x/swaprouter/types"
 )
 
 const (
@@ -14,8 +13,8 @@ const (
 )
 
 var (
-	_ sdk.Msg                       = &MsgCreateStableswapPool{}
-	_ swaproutertypes.CreatePoolMsg = &MsgCreateStableswapPool{}
+	_ sdk.Msg             = &MsgCreateStableswapPool{}
+	_ types.CreatePoolMsg = &MsgCreateStableswapPool{}
 )
 
 func NewMsgCreateStableswapPool(
@@ -117,10 +116,6 @@ func (msg MsgCreateStableswapPool) CreatePool(ctx sdk.Context, poolId uint64) (t
 	}
 
 	return &stableswapPool, nil
-}
-
-func (msg MsgCreateStableswapPool) GetPoolType() swaproutertypes.PoolType {
-	return swaproutertypes.StableSwap
 }
 
 var _ sdk.Msg = &MsgStableSwapAdjustScalingFactors{}
