+++ x/gamm/pool-models/stableswap/pool.go
@@ -11,12 +11,9 @@ import (
 	"github.com/osmosis-labs/osmosis/v13/osmomath"
 	"github.com/osmosis-labs/osmosis/v13/x/gamm/pool-models/internal/cfmm_common"
 	"github.com/osmosis-labs/osmosis/v13/x/gamm/types"
-	swaproutertypes "github.com/osmosis-labs/osmosis/v13/x/swaprouter/types"
 )
 
-var (
-	_ types.PoolI = &Pool{}
-)
+var _ types.PoolI = &Pool{}
 
 type unsortedPoolLiqError struct {
 	ActualLiquidity sdk.Coins
@@ -461,9 +458,9 @@ func validatePoolLiquidity(liquidity sdk.Coins, scalingFactors []uint64) error {
 		return liquidityAndScalingFactorCountMismatchError{LiquidityCount: liquidityCount, ScalingFactorCount: scalingFactorCount}
 	}
 
-	if liquidityCount < swaproutertypes.MinPoolAssets {
+	if liquidityCount < types.MinPoolAssets {
 		return types.ErrTooFewPoolAssets
-	} else if liquidityCount > swaproutertypes.MaxPoolAssets {
+	} else if liquidityCount > types.MaxPoolAssets {
 		return types.ErrTooManyPoolAssets
 	}
