--- x/swaprouter/client/cli/tx.go
@@ -1,360 +1,0 @@
-package cli
-
-import (
-	"errors"
-	"fmt"
-	"strconv"
-	"strings"
-	"time"
-
-	flag "github.com/spf13/pflag"
-
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/cosmos/cosmos-sdk/client/tx"
-	"github.com/spf13/cobra"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-
-	"github.com/osmosis-labs/osmosis/v13/x/gamm/pool-models/balancer"
-	"github.com/osmosis-labs/osmosis/v13/x/gamm/pool-models/stableswap"
-	"github.com/osmosis-labs/osmosis/v13/x/swaprouter/types"
-)
-
-func NewTxCmd() *cobra.Command {
-	txCmd := &cobra.Command{
-		Use:                        types.ModuleName,
-		Short:                      "Swap transaction subcommands",
-		DisableFlagParsing:         true,
-		SuggestionsMinimumDistance: 2,
-		RunE:                       client.ValidateCmd,
-	}
-
-	txCmd.AddCommand(
-		NewCreatePoolCmd(),
-		NewSwapExactAmountInCmd(),
-		NewSwapExactAmountOutCmd(),
-	)
-
-	return txCmd
-}
-
-func NewSwapExactAmountInCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "swap-exact-amount-in [token-in] [token-out-min-amount]",
-		Short: "swap exact amount in",
-		Args:  cobra.ExactArgs(2),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx, err := client.GetClientTxContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			txf := tx.NewFactoryCLI(clientCtx, cmd.Flags()).WithTxConfig(clientCtx.TxConfig).WithAccountRetriever(clientCtx.AccountRetriever)
-
-			txf, msg, err := NewBuildSwapExactAmountInMsg(clientCtx, args[0], args[1], txf, cmd.Flags())
-			if err != nil {
-				return err
-			}
-
-			return tx.GenerateOrBroadcastTxWithFactory(clientCtx, txf, msg)
-		},
-	}
-
-	cmd.Flags().AddFlagSet(FlagSetQuerySwapRoutes())
-	flags.AddTxFlagsToCmd(cmd)
-	_ = cmd.MarkFlagRequired(FlagSwapRoutePoolIds)
-	_ = cmd.MarkFlagRequired(FlagSwapRouteDenoms)
-
-	return cmd
-}
-
-func NewSwapExactAmountOutCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "swap-exact-amount-out [token-out] [token-in-max-amount]",
-		Short: "swap exact amount out",
-		Args:  cobra.ExactArgs(2),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx, err := client.GetClientTxContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			txf := tx.NewFactoryCLI(clientCtx, cmd.Flags()).WithTxConfig(clientCtx.TxConfig).WithAccountRetriever(clientCtx.AccountRetriever)
-
-			txf, msg, err := NewBuildSwapExactAmountOutMsg(clientCtx, args[0], args[1], txf, cmd.Flags())
-			if err != nil {
-				return err
-			}
-
-			return tx.GenerateOrBroadcastTxWithFactory(clientCtx, txf, msg)
-		},
-	}
-
-	cmd.Flags().AddFlagSet(FlagSetSwapAmountOutRoutes())
-	flags.AddTxFlagsToCmd(cmd)
-	_ = cmd.MarkFlagRequired(FlagSwapRoutePoolIds)
-	_ = cmd.MarkFlagRequired(FlagSwapRouteDenoms)
-
-	return cmd
-}
-
-func NewBuildSwapExactAmountInMsg(clientCtx client.Context, tokenInStr, tokenOutMinAmtStr string, txf tx.Factory, fs *flag.FlagSet) (tx.Factory, sdk.Msg, error) {
-	routes, err := swapAmountInRoutes(fs)
-	if err != nil {
-		return txf, nil, err
-	}
-
-	tokenIn, err := sdk.ParseCoinNormalized(tokenInStr)
-	if err != nil {
-		return txf, nil, err
-	}
-
-	tokenOutMinAmt, ok := sdk.NewIntFromString(tokenOutMinAmtStr)
-	if !ok {
-		return txf, nil, fmt.Errorf("invalid token out min amount, %s", tokenOutMinAmtStr)
-	}
-	msg := &types.MsgSwapExactAmountIn{
-		Sender:            clientCtx.GetFromAddress().String(),
-		Routes:            routes,
-		TokenIn:           tokenIn,
-		TokenOutMinAmount: tokenOutMinAmt,
-	}
-
-	return txf, msg, nil
-}
-
-func NewBuildSwapExactAmountOutMsg(clientCtx client.Context, tokenOutStr, tokenInMaxAmountStr string, txf tx.Factory, fs *flag.FlagSet) (tx.Factory, sdk.Msg, error) {
-	routes, err := swapAmountOutRoutes(fs)
-	if err != nil {
-		return txf, nil, err
-	}
-
-	tokenOut, err := sdk.ParseCoinNormalized(tokenOutStr)
-	if err != nil {
-		return txf, nil, err
-	}
-
-	tokenInMaxAmount, ok := sdk.NewIntFromString(tokenInMaxAmountStr)
-	if !ok {
-		return txf, nil, errors.New("invalid token in max amount")
-	}
-	msg := &types.MsgSwapExactAmountOut{
-		Sender:           clientCtx.GetFromAddress().String(),
-		Routes:           routes,
-		TokenInMaxAmount: tokenInMaxAmount,
-		TokenOut:         tokenOut,
-	}
-
-	return txf, msg, nil
-}
-
-func NewCreatePoolCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "create-pool [flags]",
-		Short: "create a new pool and provide the liquidity to it",
-		Long:  `Must provide path to a pool JSON file (--pool-file) describing the pool to be created`,
-		Example: `Sample pool JSON file contents:
-{
-	"weights": "4uatom,4osmo,2uakt",
-	"initial-deposit": "100uatom,5osmo,20uakt",
-	"swap-fee": "0.01",
-	"exit-fee": "0.01",
-	"future-governor": "168h"
-}
-`,
-		Args: cobra.ExactArgs(0),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx, err := client.GetClientTxContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			txf := tx.NewFactoryCLI(clientCtx, cmd.Flags()).WithTxConfig(clientCtx.TxConfig).WithAccountRetriever(clientCtx.AccountRetriever)
-
-			txf, msg, err := NewBuildCreateBalancerPoolMsg(clientCtx, txf, cmd.Flags())
-			if err != nil {
-				return err
-			}
-
-			return tx.GenerateOrBroadcastTxWithFactory(clientCtx, txf, msg)
-		},
-	}
-
-	cmd.Flags().AddFlagSet(FlagSetCreatePool())
-	flags.AddTxFlagsToCmd(cmd)
-
-	_ = cmd.MarkFlagRequired(FlagPoolFile)
-
-	return cmd
-}
-
-func NewBuildCreateBalancerPoolMsg(clientCtx client.Context, txf tx.Factory, fs *flag.FlagSet) (tx.Factory, sdk.Msg, error) {
-	pool, err := parseCreateBalancerPoolFlags(fs)
-	if err != nil {
-		return txf, nil, fmt.Errorf("failed to parse pool: %w", err)
-	}
-
-	deposit, err := sdk.ParseCoinsNormalized(pool.InitialDeposit)
-	if err != nil {
-		return txf, nil, err
-	}
-
-	poolAssetCoins, err := sdk.ParseDecCoins(pool.Weights)
-	if err != nil {
-		return txf, nil, err
-	}
-
-	if len(deposit) != len(poolAssetCoins) {
-		return txf, nil, errors.New("deposit tokens and token weights should have same length")
-	}
-
-	swapFee, err := sdk.NewDecFromStr(pool.SwapFee)
-	if err != nil {
-		return txf, nil, err
-	}
-
-	exitFee, err := sdk.NewDecFromStr(pool.ExitFee)
-	if err != nil {
-		return txf, nil, err
-	}
-
-	var poolAssets []balancer.PoolAsset
-	for i := 0; i < len(poolAssetCoins); i++ {
-		if poolAssetCoins[i].Denom != deposit[i].Denom {
-			return txf, nil, errors.New("deposit tokens and token weights should have same denom order")
-		}
-
-		poolAssets = append(poolAssets, balancer.PoolAsset{
-			Weight: poolAssetCoins[i].Amount.RoundInt(),
-			Token:  deposit[i],
-		})
-	}
-
-	poolParams := &balancer.PoolParams{
-		SwapFee: swapFee,
-		ExitFee: exitFee,
-	}
-
-	msg := &balancer.MsgCreateBalancerPool{
-		Sender:             clientCtx.GetFromAddress().String(),
-		PoolParams:         poolParams,
-		PoolAssets:         poolAssets,
-		FuturePoolGovernor: pool.FutureGovernor,
-	}
-
-	if (pool.SmoothWeightChangeParams != smoothWeightChangeParamsInputs{}) {
-		duration, err := time.ParseDuration(pool.SmoothWeightChangeParams.Duration)
-		if err != nil {
-			return txf, nil, fmt.Errorf("could not parse duration: %w", err)
-		}
-
-		targetPoolAssetCoins, err := sdk.ParseDecCoins(pool.SmoothWeightChangeParams.TargetPoolWeights)
-		if err != nil {
-			return txf, nil, err
-		}
-
-		var targetPoolAssets []balancer.PoolAsset
-		for i := 0; i < len(targetPoolAssetCoins); i++ {
-			if targetPoolAssetCoins[i].Denom != poolAssetCoins[i].Denom {
-				return txf, nil, errors.New("initial pool weights and target pool weights should have same denom order")
-			}
-
-			targetPoolAssets = append(targetPoolAssets, balancer.PoolAsset{
-				Weight: targetPoolAssetCoins[i].Amount.RoundInt(),
-				Token:  deposit[i],
-				// TODO: This doesn't make sense. Should only use denom, not an sdk.Coin
-			})
-		}
-
-		smoothWeightParams := balancer.SmoothWeightChangeParams{
-			Duration:           duration,
-			InitialPoolWeights: poolAssets,
-			TargetPoolWeights:  targetPoolAssets,
-		}
-
-		if pool.SmoothWeightChangeParams.StartTime != "" {
-			startTime, err := time.Parse(time.RFC3339, pool.SmoothWeightChangeParams.StartTime)
-			if err != nil {
-				return txf, nil, fmt.Errorf("could not parse time: %w", err)
-			}
-
-			smoothWeightParams.StartTime = startTime
-		}
-
-		msg.PoolParams.SmoothWeightChangeParams = &smoothWeightParams
-	}
-
-	return txf, msg, nil
-}
-
-// Apologies to whoever has to touch this next, this code is horrendous
-func NewBuildCreateStableswapPoolMsg(clientCtx client.Context, txf tx.Factory, fs *flag.FlagSet) (tx.Factory, sdk.Msg, error) {
-	flags, err := parseCreateStableswapPoolFlags(fs)
-	if err != nil {
-		return txf, nil, fmt.Errorf("failed to parse pool: %w", err)
-	}
-
-	deposit, err := ParseCoinsNoSort(flags.InitialDeposit)
-	if err != nil {
-		return txf, nil, err
-	}
-
-	swapFee, err := sdk.NewDecFromStr(flags.SwapFee)
-	if err != nil {
-		return txf, nil, err
-	}
-
-	exitFee, err := sdk.NewDecFromStr(flags.ExitFee)
-	if err != nil {
-		return txf, nil, err
-	}
-
-	poolParams := &stableswap.PoolParams{
-		SwapFee: swapFee,
-		ExitFee: exitFee,
-	}
-
-	scalingFactors := []uint64{}
-	trimmedSfString := strings.Trim(flags.ScalingFactors, "[] {}")
-	if len(trimmedSfString) > 0 {
-		ints := strings.Split(trimmedSfString, ",")
-		for _, i := range ints {
-			u, err := strconv.ParseUint(i, 10, 64)
-			if err != nil {
-				return txf, nil, err
-			}
-			scalingFactors = append(scalingFactors, u)
-		}
-		if len(scalingFactors) != len(deposit) {
-			return txf, nil, fmt.Errorf("number of scaling factors doesn't match number of assets")
-		}
-	}
-
-	msg := &stableswap.MsgCreateStableswapPool{
-		Sender:                  clientCtx.GetFromAddress().String(),
-		PoolParams:              poolParams,
-		InitialPoolLiquidity:    deposit,
-		ScalingFactors:          scalingFactors,
-		ScalingFactorController: flags.ScalingFactorController,
-		FuturePoolGovernor:      flags.FutureGovernor,
-	}
-
-	return txf, msg, nil
-}
-
-// ParseCoinsNoSort parses coins from coinsStr but does not sort them.
-// Returns error if parsing fails.
-func ParseCoinsNoSort(coinsStr string) (sdk.Coins, error) {
-	coinStrs := strings.Split(coinsStr, ",")
-	decCoins := make(sdk.DecCoins, len(coinStrs))
-	for i, coinStr := range coinStrs {
-		coin, err := sdk.ParseDecCoin(coinStr)
-		if err != nil {
-			return sdk.Coins{}, err
-		}
-
-		decCoins[i] = coin
-	}
-	return sdk.NormalizeCoins(decCoins), nil
-}
